// frontend/src/backend.d.ts
export interface Feedback {
    id: string;
    status: string;
    contactInfo: string;
    admin?: Principal;
    type: string;
    message: string;
    response?: string;
    timestamp: bigint;
}